+++ tests/test_lp.py
@@ -73,4 +73,11 @@ def test_lp():
 
     solval = myLP.solve()
 
+    assert(myLP.isOptimal())
+    assert(myLP.getPrimal() is not None)
+    assert(myLP.getDual() is not None)
+    assert(myLP.getActivity() is not None)
+    assert(myLP.getRedcost() is not None)
+    assert round(myLP.getObjVal() == solval)
+
     assert round(5.0 == solval)