+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 ### Added
 - Added getLinearConsIndicator
-- Added SCIP_LPPARAM, setIntParam, setRealParam, getIntParam, getRealParam
+- Added SCIP_LPPARAM, setIntParam, setRealParam, getIntParam, getRealParam for lpi
+- Added isFeasPositive
 ### Fixed
 ### Changed
 ### Removed