+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 ### Added
 - Added getLinearConsIndicator
-- Added SCIP_LPPARAM, setIntParam, setRealParam, getIntParam, getRealParam for lpi
+- Added SCIP_LPPARAM, setIntParam, setRealParam, getIntParam, getRealParam, isOptimal, getObjVal, getRedcost for lpi
 - Added isFeasPositive
 ### Fixed
 ### Changed