+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 ### Added
-- Wrapper SCIPdeactivatePricer, SCIPsetConsModifiable and added deactivatePricer, setModifiable and test
+- Wrapped SCIPgetSiblings, SCIPgetNSiblings, and added getSiblings, getNSiblings and test
+- Wrapped SCIPdeactivatePricer, SCIPsetConsModifiable, and added deactivatePricer, setModifiable and test
 - Added getLinearConsIndicator
 - Added SCIP_LPPARAM, setIntParam, setRealParam, getIntParam, getRealParam, isOptimal, getObjVal, getRedcost for lpi
 - Added isFeasPositive